requirements.txt
torch==1.9.0
torchvision==0.10.0

einops==0.3.0
kornia==0.5.7

Pillow==8.3.2
numpy==1.20.2

requests==2.24.0
tqdm==4.51.0

regex==2021.4.4
ftfy==6.0.3

jupyter==1.0.0
pandas==1.3.*
matplotlib==3.5.1
imageio==2.14.0
